rtqa gui-calc separation
